refactor: Room 적용 후 기존 CartDatabase 삭제 및 패키지 구조 변경
