+++ app/src/main/java/woowacourse/shopping/product/catalog/CatalogViewModelFactory.kt
@@ -4,7 +4,7 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
 import woowacourse.shopping.ShoppingApplication
 import woowacourse.shopping.data.database.ShoppingDatabase
-import woowacourse.shopping.repository.CartProductRepositoryImpl
+import woowacourse.shopping.data.repository.CartProductRepositoryImpl
 
 class CatalogViewModelFactory(
     private val application: ShoppingApplication,
